+++ lib/screen/landing.dart
@@ -9,7 +9,6 @@ import 'package:slide_puzzle/code/models.dart';
 import 'package:slide_puzzle/code/providers.dart';
 import 'package:slide_puzzle/screen/app.dart';
 import 'package:slide_puzzle/ui/3d_transform.dart';
-import 'package:slide_puzzle/ui/Scoreboard.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:slide_puzzle/ui/button.dart';
@@ -91,6 +90,7 @@ class _LandingPageState extends State<LandingPage>
     WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
       tileProvider.updateImages(context);
     });
+    // DatabaseService.instance.submitDummyCommunityScores();
   }
 
   @override
@@ -168,15 +168,15 @@ class _LandingPageState extends State<LandingPage>
                               .headline5!
                               .copyWith(color: secondaryColor),
                         ),
-                        userData != null
-                            ? ScoreBoard(
-                                gridSize: 3,
-                                // currentMove: 0,
-                                // currentTime: 0,
-                                userData: userData,
-                                child: Container(),
-                              )
-                            : Container(),
+                        // userData != null
+                        //     ? ScoreBoard(
+                        //         gridSize: 3,
+                        //         // currentMove: 0,
+                        //         // currentTime: 0,
+                        //         userData: userData,
+                        //         child: Container(),
+                        //       )
+                        //     : Container(),
                       ],
                     )),
                     Align(